refactor(25905): refactor hook into a context provider
