+++ hivemq-edge/src/frontend/src/hooks/usePrivateMqttClient/PrivateMqttClientContext.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react'
+import { PrivateMqttClientType } from '@/hooks/usePrivateMqttClient/type.ts'
+
+export const PrivateMqttClientContext = createContext<PrivateMqttClientType>({ state: undefined, actions: undefined })
