+++ hivemq-edge/src/frontend/src/hooks/usePrivateMqttClient/mqtt-client.utils.ts
@@ -1,2 +1,4 @@
+/* istanbul ignore next -- @preserve */
 export const SAMPLING_DURATION = 1000
+/* istanbul ignore next -- @preserve */
 export const PRIVATE_MQTT_CLIENT = 'ws://localhost:8000/mqtt'
