+++ hivemq-edge/src/frontend/src/hooks/usePrivateMqttClient/usePrivateMqttClient.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react'
+import { PrivateMqttClientContext } from '@/hooks/usePrivateMqttClient/PrivateMqttClientContext.ts'
+
+export const usePrivateMqttClient = () => {
+  return useContext(PrivateMqttClientContext)
+}
